+++ Deps/encrypting.hpp
@@ -17,7 +17,7 @@ struct result {
 
 namespace cp {
     std::vector<uint8_t> fixed_xor(std::vector<uint8_t> &start, std::vector<uint8_t> &key);
-    int character_lookup(std::vector<uint8_t> &byte_array, int &key);
+    int character_lookup(std::vector<uint8_t> &byte_array, uint8_t &key);
     std::map<int, uint8_t> get_scores(std::vector<uint8_t> &byte_array);
     std::vector<uint8_t> single_byte_xor(std::vector<uint8_t> &byte_array, uint8_t &key);
     result attack_single_byte_xor(std::vector<uint8_t> &cipher_bytes);
